// examples/uss-omaha_batch-4/uss-omaha/src/types/content.ts
export interface Volunteer {
  heading: string
  subheading: string
  contact: Contact
  opportunities: string[]
  organization: string
  organizationContact: string
}

export interface Contact {
  name: string
  phone: string
  email: string
}